+++ changelog.md
@@ -11,6 +11,7 @@
 * LLT-3629: Add IPv6 support to nat-lab
 * LLT-4119: Fix multiple bugs in telio-firewall. Mainly whitelist vpn server
 * LLT-4166: Bump tokio-rustls version
+* LLT-3913: Deduplicate IPs of external wireguard peers
 
 <br>